Separate reset button js_only
