+++ package.json
@@ -232,6 +232,7 @@
 		"lint-staged": "^9.4.2",
 		"lodash": "^4.17.15",
 		"mini-css-extract-plugin": "^0.8.0",
+		"node-sass": "^4.13.1",
 		"pegjs": "^0.10.0",
 		"pegjs-loader": "^0.5.5",
 		"postcss-loader": "^3.0.0",
